requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
defusedxml==0.7.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.115.12
filelock==3.18.0
fonttools==4.58.1
fpdf==1.7.2
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.2
huggingface-hub==0.32.3
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.5
nltk==3.9.1
numpy==2.2.6
packaging==25.0
pillow==11.2.1
preshed==3.0.10
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-multipart==0.0.20
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scipy==1.15.3
sentence-transformers==4.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.2
sympy==1.14.0
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.52.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.3
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
